refactor(tests): move tests to their own directory
