+++ src/server/index.ts
@@ -2,3 +2,4 @@ export * from "./loaders";
 export * from "./caches";
 export * from "./transformers";
 export * from "./resolvers";
+export * from "../utils/cache";
